data/importers/requirements.txt
numpy==1.21.1
pandas==1.3.1
geopandas==0.10.2
rasterio==1.3.6
h3ronpy==0.16.1
h3==3.7.6
psycopg2-binary==2.9.1
docopt==0.6.0
tqdm==4.62.3
jsonschema==4.6.0
opencv-python==4.6.0.66
requests
click==8.1.3
rich==13.3.1
psycopg==3.1.8
psycopg_pool==3.1.6
boto3==1.26
